ajout de ft_memcmp

+++ srcs/mem/ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: snicolet <marvin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 12:17:00 by snicolet          #+#    #+#             */
-/*   Updated: 2015/11/24 14:53:15 by snicolet         ###   ########.fr       */
+/*   Updated: 2015/11/24 14:58:41 by snicolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_memset(char *str, char c, unsigned int n)
+#include "unistd.h"
+
+void	ft_memset(char *str, char c, size_t n)
 {
 	while (n--)
 		str[n] = c;
